--- src/app/(marketplace)/shopping/components/filterSection/index.ts
@@ -1,1 +1,0 @@
-export * from './FilterSection';
